requirements.txt
torch
numpy
Pillow
tabulate
tqdm
torchvision
pyheif
pillow-heif
scikit-learn

gdown